requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
pyttsx3
langchain 
sentence-transformers  
sqlalchemy 
psycopg2
SpeechRecognition
fitz
frontend
tools
PyMuPDF
mysql-connector-python
pyodbc
redis
requests
python-dotenv
